requirements.txt
glfw
moderngl==5.6.0
pyglm
